pyproject.toml
[project]
name = "tech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangChain 生态
    "langchain-core>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    # LLM
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    # 数据验证
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # 视觉模型
    "transformers>=5.0.0",
    "ultralytics>=8.0.0",
    # 文件处理
    "pymupdf>=1.23.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
